refactor(chat): extract hub event forwarder
